pyproject.toml
[tool.poetry]
name = "parsing hh"
version = "0.1.0"
description = ""
authors = ["Kris <ssspers92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
